requirements.txt
#Project
tabulate
colorama
pylint
configparser
cryptography

#Database
pyodbc
fastapi
uvicorn[standard]
pydantic
azure-identity
